Address PR review feedback.
